+++ tests/unit/connector/test_confluent_kafka_output.py
@@ -6,7 +6,6 @@
 # pylint: disable=no-self-use
 
 import json
-from copy import deepcopy
 from unittest import mock
 
 import pytest
@@ -22,33 +21,13 @@ from tests.unit.connector.test_confluent_kafka_common import (
 class TestConfluentKafkaOutput(BaseOutputTestCase, CommonConfluentKafkaTestCase):
     CONFIG = {
         "type": "confluentkafka_output",
-        "bootstrapservers": ["testserver:9092"],
         "topic": "test_input_raw",
         "error_topic": "test_error_topic",
         "flush_timeout": 0.1,
-        "ssl": {
-            "cafile": "test_cafile",
-            "certfile": "test_certfile",
-            "keyfile": "test_keyfile",
-            "password": "test_password",
+        "kafka_config": {
+            "bootstrap.servers": "testserver:9092",
         },
     }
-
-    def test_confluent_settings_contains_expected_values(self):
-        expected_config = {
-            "bootstrap.servers": "testserver:9092",
-            "security.protocol": "SSL",
-            "ssl.ca.location": "test_cafile",
-            "ssl.certificate.location": "test_certfile",
-            "ssl.key.location": "test_keyfile",
-            "ssl.key.password": "test_password",
-            "queue.buffering.max.messages": 100000,
-            "compression.type": "none",
-            "acks": -1,
-            "linger.ms": 0.5,
-        }
-        kafka_input_cfg = self.object._confluent_settings
-        assert kafka_input_cfg == expected_config
 
     @mock.patch("logprep.connector.confluent_kafka.output.Producer", return_value="The Producer")
     def test_producer_property_instanciates_kafka_producer(self, _):
@@ -144,15 +123,6 @@ class TestConfluentKafkaOutput(BaseOutputTestCase, CommonConfluentKafkaTestCase)
         self.object.input_connector = mock.MagicMock()
         self.object.store({"message": "my event message"})
         self.object.input_connector.batch_finished_callback.assert_called()
-
-    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
-    def test_logprep_config_has_precedence(self, mock_producer):
-        kafka_config = deepcopy(self.object._confluent_settings)
-        config = {"bootstrap.servers": "bootstrap1, myprivatebootstrap"}
-        self.object._config.kafka_config = config
-        self.object._producer.clear()
-        _ = self.object._producer
-        mock_producer.assert_called_with(kafka_config)
 
     def test_setup_raises_fatal_output_error_on_invalid_config(self):
         config = {"myconfig": "the config"}