+++ CHANGELOG.md
@@ -18,7 +18,8 @@ It contains the time at which a pre-detection was created by the processor.
 * add `prometheus` and `grafana` to the quickstart setup to support development
 * reimplemented kafka input connector
   - move kafka config options to `kafka_config` dictionary
-  - wait until poll returns message (never return `None` in `get_next`)
+* reimplement kafka output connector
+  - move kafka config options to `kafka_config` dictionary
 
 ### Bugfix